Elijah's Stuff/requirements.txt
flask
wtforms
Flask-WTF

# Roboflow Inference
inference-gpu
roboflow

# Tesseract OCR Extra Feature
tesseract
pytesseract
Pillow